% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C.ATS.R
\name{C.ATS}
\alias{C.ATS}
\title{A Cat Function -  To test out personal packages and functions}
\usage{
C.ATS(love = TRUE)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function tests your love of catsI
}
\details{
When building your package, remember to make the documentation
devtools::document()
}
\examples{
C.ATS()
}
\keyword{cats}
